The downgrade to hugo v0.55.6 so that single hugo-fresh pages work.
The "single" pages do not seem to compile correctly when the hugo-fresh theme
is compiled using the latest hugo version from homebrew (0.57.2).

Downgrading hugo to version 0.55.6 seems to do the trick locally, let's see
if it works on netlify with updated environment variables.

Checkout
[this github issue](https://github.com/StefMa/hugo-fresh/issues/70)
for more details and to follow progress resolving this issue.